financial_ai_agent_v2/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
zhipuai>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
sniffio>=1.3.0
httpx>=0.24.0
anyio>=3.7.0
pydantic>=2.0.0
streamlit-authenticator>=0.2.3
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
